pagespec_alias: sanity-check pagespec aliases
Since user-supplied pagespec-aliases are turned into bits
of dynamic Perl code, they could be abused.  Ensure that
all user-supplied pagespec alias names are composed soley
of word-characters (\w+).  Ignore any that aren't.

I considered calling error() if a bad pagespec alias was
detected, but this would mean web-based ikiwikis would be
un-editable until the bad alias was removed, a potential
DoS.  On the other hand,  the bad alias would be guaranteed
not to be noticed, and if someone naughty has admin access
to the web interface, there are already problems.