pagespec_aliases: rework getsetup
getsetup had placeholder values.  Replace them with more sensible
choices for the plugin.  In particular, mark it as unsafe for now
(which will hide it from websetup.  It was already hidden by
virtue of using an unsupported data structure…)

It isn't possible to express the "options" in getsetup() properly
yet. We fold a hash structure into a string, which gives you
something *nearly* right as an example, which is as best as I can
do without enhancing the getsetup code in core.