Make environment keys preserved by wrapper controllable by plugins.
First step of a refactoring so that a plugin can influence the set of
environment keys that will be preserved by a generated wrapper, rather
than having Wrapper.pm hardcode, for example, what environment keys are
needed by CGI.pm and Plugin/404.pm.

In this first step, nothing happens except the same hardcoded list of
keys for CGI is moved from Wrapper.pm to CGI.pm, and the code that adds
the CGI wrapper spec to %config (which happens to be in Setup.pm) now
inserts a 'needenvkeys' key that supplies them.

This change does seem to have exposed a thing where various plugins that
call checksessionexpiry() (in CGI.pm) have been supplying one more argument
than its prototype allows ... for years ... apparently because the prototype
hadn't been seen at compilation time for those source files, so no error was
reported. For now, instead of fixing all those call sites, I've changed the
prototype to allow an optional third argument that's completely ignored (which
is what was happening before anyway).